added log.js  backend using the getlogentry function them implemented it in camera.tsx
